final commit

+++ app.py
@@ -1,6 +1,6 @@
 from flask import Flask, request, jsonify
 from flask_sqlalchemy import SQLAlchemy
-
+from datetime import datetime
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://mariadbtest:mariadbtest@mariadb114.r5.websupport.sk:3306/mariadbtest'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
@@ -21,6 +21,7 @@ def customers_all():
         customers_dict.append(customer.dict())
     return jsonify(customers_dict)
 
+# POSTMAN http://127.0.0.1:5000/orders
 @app.route("/orders")
 def orders_all():
     orders = Order.query.all()
@@ -29,23 +30,27 @@ def orders_all():
         orders_dict.append(order.dict())
     return jsonify(orders_dict)
 
+#POSTMAN: http://127.0.0.1:5000/customers/2/orders
 @app.route("/customers/<customer_id>/orders")
 def orders_by_customer(customer_id):
     orders = Order.query.filter_by(customer_id=customer_id).all()
     orders_dict = []
     for order in orders:
         orders_dict.append(order.dict())
-    return jsonify(orders.dict)
+    return jsonify(orders_dict)
 
+#POSTMAN: http://127.0.0.1:5000/customers/1/orders (insert –> raw / JSON: product_name, quantity)
 @app.route("/customers/<customer_id>/orders", methods=['POST'])
 def create_order(customer_id):
     order = Order(customer_id=customer_id,
-                  product_id=request.json['product_id'],
-                  quantity=request.json['quantity'])
+                  product_name=request.json['product_name'],
+                  quantity=request.json['quantity'],
+                  order_date=datetime.utcnow())
     db.session.add(order)
     db.session.commit()
     return jsonify(order.dict())
 
+#POSTMAN: http://127.0.0.1:5000/orders/1 JSON: quantity
 @app.route("/orders/<order_id>", methods=['PUT'])
 def update_order(order_id):
     order = Order.query.get(order_id)
@@ -56,7 +61,12 @@ def update_order(order_id):
 @app.route("/orders/<order_id>", methods=['DELETE'])
 def delete_order(order_id):
     order = Order.query.get(order_id)
+    if not order:
+        return jsonify({'error': 'Order not found'}), 404
+
     db.session.delete(order)
+    db.session.commit()
+    return jsonify({'message': 'Order deleted successfully'})
 
 if __name__ == '__main__':
     app.run()
